+++ benches/fibonacci.rs
@@ -8,7 +8,6 @@ use std::{sync::Arc, time::Duration};
 
 use lurk::{
     eval::lang::{Coproc, Lang},
-    fib::{fib_expr, fib_frame, fib_limit},
     lem::{eval::evaluate, multiframe::MultiFrame, store::Store},
     proof::nova::NovaProver,
     proof::Prover,
@@ -19,7 +18,12 @@ use lurk::{
 };
 
 mod common;
-use common::set_bench_config;
+use common::{
+    fib::{fib_expr, fib_frame, fib_limit},
+    set_bench_config,
+};
+
+use crate::common::fib::{test_coeffs, test_fib_io_matches};
 
 #[derive(Clone, Debug, Copy)]
 struct ProveParams {
@@ -126,9 +130,14 @@ fn fibonacci_prove<M: measurement::Measurement>(
 }
 
 fn fibonacci_benchmark(c: &mut Criterion) {
+    // Running tests to make sure the constants are correct
+    test_coeffs();
+    test_fib_io_matches();
+
     // Uncomment to record the logs. May negatively impact performance
     //tracing_subscriber::fmt::init();
     set_bench_config();
+
     tracing::debug!("{:?}", lurk::config::LURK_CONFIG);
 
     let reduction_counts = rc_env().unwrap_or_else(|_| vec![100]);
